Update libbpf-dev package installation
Changed the version of the libbpf-dev package to be installed by dpkg in the GitHub dynamic workflow script. Also removed the specific filename from the dpkg installation command and replaced it with a trick to select the recently downloaded package, to make it more flexible and robust against version changes.