Refactor dependencies installation in GitHub workflow
The linux-tools-common has been removed from the list of dependencies to be installed, aligning with the goal of reducing superfluous dependencies. Additionally, the commit introduces the cloning of the libbpf/bpftool repository followed by installing from its source. These changes aim to simplify the dependency installation process, enhancing the efficiency of the GitHub workflow runs.