Add BPFTOOL_VERSION variable in GitHub workflow
With this update, the environment variable BPFTOOL_VERSION is added to the GitHub build workflow. Instead of installing libbpf/bpftool from source, the specific BPFTOOL_VERSION is now downloaded as a tar.gz file. This makes the workflow more efficient by ensuring a specific version of bpftool is used consistently across builds.