Add 'staty' struct to core package
This commit adds a new 'staty' struct to the core package which consolidates FQDN, IP, protocol, and port information. A new method 'String' is also introduced for this struct that returns a formatted string representation of 'staty'. The new struct will help in identifying and organizing assets in the Eggy struct.